Use IGRAB-compatible names for graphics chunks
Bio Menace's graphics were originally imported using a tool called
IGRAB, which generates the GRAPHBM1.H and GRAPHBM1.EQU files with the
chunk numbers, and other metadata.

IGRAB, however, has a number of rules when generating these names:
- Bitmaps have 'PIC' appended to the name.
- Masked bitmaps have 'PICM' appended to the name.
- Sprites have 'SPR' appended to the name.
- (Other chunks have no naming restrictions)

The Bio Menace decomp uses a handwritten GRAPHBM1.{H,EQU} file, as the
originals are not available (and neither is the IGRAB .I script).
However, the names used do not conform to the original IGRAB rules, so
it would not be possible to reconstruct such a file, or to use other
tools which expect such formatting, to create compatible headers.

In particular:
- The 'ceilwalker' graphics chunks are missing a 'SPR' suffix.
- Where a 'SPR' suffix is used, it is often (but not always) separated
  with an underscore, making it '_SPR'. This is not a problem, but does
  look a bit weird (IGRAB must be given a name with a trailing
  underscore).
- The PIC and PICM suffixes are not separated by an underscore, which is
  inconsistent.

This change mechanically adds '_SPR' as a suffix where it was not
present, and adds an underscore where one was missing (for PIC, PICM,
and a few sprites like SCOREBOXSPR). This provides consistency, and is
technically compatible with IGRAB, so long as all of the bitmap, masked
bitmap, and sprite names end in a trailing underscore.

I've tested this with headers generated from my own in-progress 'idGrab'
tool, which can generate IGRAB-style headers, (and can automatically add
an underscore as a separator), and the game builds and runs fine.